fix: add checks for delete_message endpoint

+++ backend/src/api/chat/routes.py
@@ -80,7 +80,10 @@ async def get_messages(dialog_id: int, amount: int = 0) -> list[ViewMessage]:
 
 @router.delete("/chat/delete_message")
 async def delete_message(message_id: int) -> None:
-    await messages_repository.delete_message(message_id)
+    message = await messages_repository.get_message_by_id(message_id)
+    if message.role != Roles.USER:
+        raise HTTPException(400, "You can only delete messages from user")
+    return await messages_repository.delete_message(message_id)
 
 
 @router.post("/chat/regenerate")
